backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
aiofiles==24.1.0

# Document parsing
pdfplumber==0.11.4
openpyxl==3.1.5
pandas==2.2.3
ezdxf==1.3.4
Pillow==10.4.0
pytesseract==0.3.13
